Cleaning up Ui functions
