+++ Main.py
@@ -1,0 +1,4 @@
+import UIAutomationFunctions
+import time
+
+UIAutomationFunctions.open_application('ms-settings:')
